+++ src/_utils/checkboxIcon/index.tsx
@@ -8,16 +8,28 @@ import { color } from '../branding'
 interface CheckboxIconProps {
   readonly isChecked?: boolean
   readonly isLoading?: boolean
+  readonly isDisabled?: boolean
 }
 
-const CheckboxIcon = ({ isChecked = false, isLoading = false }: CheckboxIconProps) => {
+const CheckboxIcon = ({
+  isChecked = false,
+  isLoading = false,
+  isDisabled = false,
+}: CheckboxIconProps) => {
   if (isLoading) {
     return <Loader size={24} inline />
   }
   if (isChecked) {
-    return <CheckIcon iconColor={color.white} backgroundColor={color.blue} thin />
+    return (
+      <CheckIcon
+        backgroundColor={color.blue}
+        iconColor={color.white}
+        isDisabled={isDisabled}
+        thin
+      />
+    )
   }
-  return <CircleIcon iconColor={color.blue} />
+  return <CircleIcon iconColor={color.blue} isDisabled={isDisabled} />
 }
 
 export default CheckboxIcon