+++ src/icon/checkIcon.tsx
@@ -1,8 +1,9 @@
 // tslint:disable:max-line-length
-import React, { Fragment } from 'react'
+import React from 'react'
 import cc from 'classcat'
 import styled from 'styled-components'
 
+import { color } from '../_utils/branding'
 import BaseIcon from '../_utils/icon'
 import { BaseIconDefaultProps } from '../_utils/icon/BaseIcon'
 
@@ -12,6 +13,8 @@ export interface CheckProps extends Icon {
   readonly backgroundColor?: string
   readonly thin?: boolean
 }
+
+const defaultBackgroundColor = 'transparent'
 
 const CheckIcon = ({ absolute, validate, backgroundColor, thin, ...props }: CheckProps) => (
   <BaseIcon
@@ -25,37 +28,55 @@ const CheckIcon = ({ absolute, validate, backgroundColor, thin, ...props }: Chec
       },
     ])}
   >
-    <Fragment>
-      <path
-        d="M6.5 12.5l4 4 8-8"
-        fill="none"
-        stroke={props.iconColor}
-        strokeWidth={thin ? '1' : '2'}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeMiterlimit="10"
-      />
-    </Fragment>
+    <path
+      d="M6.5 12.5l4 4 8-8"
+      fill="none"
+      strokeWidth={thin ? '1' : '2'}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeMiterlimit="10"
+    />
   </BaseIcon>
 )
 
 export const StyledCheckIcon = styled(CheckIcon)`
   & {
-    background-color: ${props => props.backgroundColor};
+    background-color: ${props =>
+      props.isDisabled && props.backgroundColor !== defaultBackgroundColor
+        ? color.gray
+        : props.backgroundColor};
     border-radius: 100%;
   }
+
+  & path {
+    stroke: ${props => {
+      // If there's an outer background circle, filled,
+      // then when disabled it's the background that turns gray and the check is white.
+      // If not (transparent default background), we gray out the check (path) when disabled.
+      if (props.isDisabled && props.backgroundColor !== defaultBackgroundColor) {
+        return color.white
+      }
+      if (props.isDisabled) {
+        return color.gray
+      }
+      return props.iconColor
+    }};
+  }
+
   &.absolute {
     position: absolute;
     left: 50%;
     top: 50%;
     transform: translateX(-50%) translateY(-50%);
   }
+
   &.validate path {
     stroke-dasharray: 24;
     stroke-dashoffset: 24;
     stroke-linecap: round;
     animation: dash 0.5s cubic-bezier(0.65, 0, 0.45, 1) forwards;
   }
+
   @keyframes dash {
     from {
       stroke-dashoffset: 24;
@@ -70,7 +91,7 @@ StyledCheckIcon.defaultProps = {
   ...BaseIconDefaultProps,
   absolute: false,
   validate: false,
-  backgroundColor: 'transparent',
+  backgroundColor: defaultBackgroundColor,
   thin: false,
 }