+++ src/itemChoice/story.tsx
@@ -3,7 +3,8 @@ import { action } from '@storybook/addon-actions'
 import { boolean, select, text, withKnobs } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
 
-import ComfortIcon from '../icon/comfortIcon'
+import CheckIcon from '../icon/checkIcon'
+import MyRidesIcon from '../icon/myRides'
 import ItemChoice, { ItemChoiceStatus, ItemChoiceStyle } from '../itemChoice'
 import Section from '../layout/section/baseSection'
 import specs from './specifications/index.md'
@@ -17,6 +18,7 @@ stories.add(
     const isLink = boolean('Use as link', false)
     const withLeftIcon = boolean('With left icon', false)
     const withRightIcon = boolean('With right icon', false)
+    const disabled = boolean('Disabled', false)
 
     return (
       <Section>
@@ -25,14 +27,14 @@ stories.add(
           labelInfo={text('Title Info', 'Additional informations')}
           data={text('Data', 'Data')}
           dataInfo={text('Data Info', 'Info')}
-          leftAddon={withLeftIcon ? <ComfortIcon /> : null}
-          rightAddon={withRightIcon ? <ComfortIcon /> : null}
+          leftAddon={withLeftIcon ? <CheckIcon isDisabled={disabled} /> : null}
+          rightAddon={withRightIcon ? <MyRidesIcon active isDisabled={disabled} /> : null}
           status={select('Status', ItemChoiceStatus, ItemChoiceStatus.DEFAULT)}
           style={select('Style', ItemChoiceStyle, ItemChoiceStyle.PRIMARY)}
           onDoneAnimationEnd={action('onDoneAnimationEnd')}
           onClick={action('onClick')}
           href={isLink ? <a href="#" /> : ''}
-          disabled={boolean('Disabled', false)}
+          disabled={disabled}
           ariaLabel={text('Aria label', 'Aria label')}
         />
       </Section>