+++ src/itemInfo/story.tsx
@@ -10,14 +10,13 @@ const stories = storiesOf('Widgets|ItemInfo', module)
 stories.addDecorator(withKnobs)
 
 stories.add('ItemInfo', () => {
-  const hasMainTitle = boolean('Title', true)
   const hasIcon = boolean('Icon', true)
 
   return (
     <Section>
       <ItemInfo
+        mainTitle={text('Main title')}
         mainInfo={text('Main info', 'Main information')}
-        mainTitle={hasMainTitle ? text('Main title', 'Title') : null}
         icon={hasIcon ? <ClockIcon /> : null}
         tag={<li />}
         ariaLabel={text('Aria Label', 'Custom aria label')}